Add `JITExecManager` for coordinated execution of JIT-compiled functions
- Introduced `JITExecManager` to manage queued and immediate execution strategies for JIT-compiled functions.
- Integrated newly added manager into the `CompilerInstance`, enabling seamless submission and execution of JIT tasks.
- Enhanced Sema to coordinate with `JITExecManager` for executing statements and expressions, simplifying function call handling.
- Updated build and project files to include `JITExecManager` sources and headers.

+++ src/Compiler/Sema/Match/Match.Func.cpp
@@ -1037,7 +1037,7 @@ namespace
         }
 
         SmallVector4<uint64_t> dims;
-        dims.push_back(static_cast<uint64_t>(elemTypes.size()));
+        dims.push_back(elemTypes.size());
         const TypeRef concreteArrayTypeRef = sema.typeMgr().addType(TypeInfo::makeArray(dims, firstElemTypeRef));
 
         SemaNodeView castView(sema, argRef, SemaNodeViewPartE::Node | SemaNodeViewPartE::Type | SemaNodeViewPartE::Constant);
